currency type: basic IO
The currency type combines the NUMERIC type with the TLA type, for a
currency value and a currency code.  Like the taggedtypes example, it
allocates new buffers for all calls to the numeric IO functions.  This
saves 4 bytes (eg, 14 instead of 18 for '100 EUR') per value versus just
wrapping the entire varlena structure.